+++ packages/jaspr/lib/src/server/adapters/client_component_adapter.dart
@@ -1,6 +1,5 @@
 import '../../foundation/options.dart';
 import '../../framework/framework.dart';
-import '../child_nodes.dart';
 import '../markup_render_object.dart';
 import '../server_binding.dart';
 import 'client_script_adapter.dart';